requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.4
argcomplete==1.12.2
attrs==20.3.0
black==20.8b1
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorlog==4.6.2
coverage==5.3
cryptography==3.3.1
distlib==0.3.1
docutils==0.16
filelock==3.0.12
flake8==3.8.4
flake8-copyright==0.2.2
future==0.18.2
-e git+https://github.com/facebookresearch/hydra/@85f7b18da5f8421cb34f703888e0b45fdab0c9e4#egg=hydra_configen&subdirectory=tools/configen
-e git+git@github.com:pytorch/hydra-torch.git@2a8e7f4fbff0b5bf35972f02208f35e29e56163c#egg=hydra_configs_torch&subdirectory=hydra-configs-torch
hydra-core==1.0.4
identify==1.5.10
idna==2.10
importlib-resources==3.3.0
incremental==17.5.0
iniconfig==1.1.1
isort==5.5.2
jeepney==0.6.0
Jinja2==2.11.2
keyring==21.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
nodeenv==1.5.0
nox==2020.8.22
numpy==1.19.4
omegaconf==2.0.5
packaging==20.8
pathspec==0.8.1
pkginfo==1.6.1
pluggy==0.13.1
pre-commit==2.9.3
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.3
pyparsing==2.4.7
pytest==6.2.1
pytest-snail==0.1.0
PyYAML==5.3.1
readme-renderer==28.0
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
SecretStorage==3.3.0
six==1.15.0
toml==0.10.2
torch==1.6.0
towncrier==19.2.0
tqdm==4.54.1
twine==3.2.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
webencodings==0.5.1
yamllint==1.25.0
